fix grpc health checks closing conn after single failure
A change ~2 years ago seems to have accidentally broken grpc health
checks and resulted in the conn getting closed after a single
healthcheck returns an error.

A check that was meant to see if the outer-scope's ctx was canceled
instead checked a different ctx from the inner scope that is hardcoded
always canceled. This resulted in `monitorHealth` returning after a
single healthcheck failure and closing the connection. This is likely
one of the causes of "session not found" errors in dagger, based on
observing logs where the session attachable conn got closed exactly 20s
after it was started, with 20s being the default grpc connection
timeout.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>